refactor(ai-tools): extract 6 ticket convenience handlers to operation-handlers/ticket-convenience
